backend/microservices/user_management/requirements.txt
alembic==1.13.1
docker==7.1.0
grpc-interceptor==0.15.4
grpcio==1.64.1
grpcio-tools==1.64.1
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
SQLAlchemy==2.0.30
uvloop==0.19.0